// src/game/map/generatedShipMapLayout.ts
export const SHIP_MAP_LAYOUT = {
// ...
    {
      id: "level-five-shuttle-outline",
      levelId: "level-five",
      type: "path",
      x: 1712,
      y: 8292,
      width: 1876,
      height: 523,
      points: [{ x: 1712, y: 8471 }, { x: 1794, y: 8451 }, { x: 1794, y: 8433 }, { x: 1862, y: 8417 }, { x: 1964, y: 8292 }, { x: 2098, y: 8296 }, { x: 2464, y: 8443 }, { x: 2910, y: 8475 }, { x: 3290, y: 8549 }, { x: 3508, y: 8633 }, { x: 3588, y: 8675 }, { x: 3588, y: 8705 }, { x: 3512, y: 8749 }, { x: 3292, y: 8785 }, { x: 2910, y: 8815 }, { x: 2464, y: 8799 }, { x: 2098, y: 8739 }, { x: 1964, y: 8751 }, { x: 1862, y: 8607 }, { x: 1794, y: 8595 }, { x: 1794, y: 8573 }, { x: 1712, y: 8561 }, { x: 1712, y: 8537 }, { x: 1794, y: 8523 }, { x: 1794, y: 8499 }, { x: 1712, y: 8493 }],
      closed: true,
      roomIds: ["InsideShuttle", "ShuttleCockpit"],
    }
// ...
} as const satisfies ShipMapLayout;